Apply rust review batch 2: status codes, spin loops, and lookups
- health endpoint returns 500 with log detail instead of 200 body
- build coordinator breaks on broadcast Closed instead of warn-spin
- version check indexes RPC results by pkgbase (was O(PxR) scan)
- sort_unstable for snapshot/describe listings, whose keys are unique;
  AUR popularity keeps the stable sort, since ties are common there
- dashboard success_rate adds wide (u32 wrap at 2^32 builds)

+++ frontend-rs/src/screens/dashboard.rs
@@ -84,8 +84,10 @@ pub fn Dashboard() -> Element {
 /// would make a busy queue look like a broken server. `None` when nothing has
 /// finished, because zero out of zero is not zero per cent.
 fn success_rate(successful: u32, failed: u32) -> Option<f64> {
-    let finished = successful + failed;
-    (finished > 0).then(|| f64::from(successful) / f64::from(finished) * 100.0)
+    // Added wide: the counters run over an unbounded history, so `u32 + u32`
+    // would wrap at 2^32 builds into a wrong percentage.
+    let finished = u64::from(successful) + u64::from(failed);
+    (finished > 0).then(|| u64::from(successful) as f64 / finished as f64 * 100.0)
 }
 
 /// A rate as a percentage, or a dash.
